+++ R/ggseqfplot.R
@@ -47,11 +47,7 @@
 #' \code{\link[ggseqplot:ggseqiplot]{ggseqiplot}}
 #'
 #' @examples
-#' # Use examples from TraMineR & load ggplot for fine-tuning the plot
-#' library(TraMineR)
-#' library(ggplot2)
-#'
-#' # actcal data set
+#' # Use example data from TraMineR: actcal data set
 #' data(actcal)
 #'
 #' # We use only a sample of 300 cases