+++ R/ggseqmtplot.R
@@ -45,11 +45,7 @@
 #'   \insertAllCited{}
 #'
 #' @examples
-#' # Use examples from TraMineR & load ggplot for fine-tuning the plot
-#' library(TraMineR)
-#' library(ggplot2)
-#'
-#' # actcal data set
+#' # Use example data from TraMineR: actcal data set
 #' data(actcal)
 #'
 #' # We use only a sample of 300 cases